Add tests and docker

+++ src/main/java/com/example/servingwebcontent/service/dto/DeviceCreateDto.java
@@ -14,5 +14,5 @@ public class DeviceCreateDto {
     String name;
     @NotNull
     Type type;
-    Integer ownerId;
+    Long ownerId;
 }
