+++ src/main/java/com/example/servingwebcontent/service/dto/DeviceDto.java
@@ -16,5 +16,5 @@ public class DeviceDto {
     String name;
     @NotNull
     Type type;
-    Integer ownerId;
+    Long ownerId;
 }
